refactor: add type for password, remove console.log, change ctyle file form .css to .scss, and add interpolation
